Rename challenges to questions
This keeps the naming between frontend and the backend consistent
