+++ lingetic-nextjs-frontend/__tests__/components/challenges/FillInTheBlanks/FillInTheBlanks.test.tsx
@@ -6,7 +6,7 @@ import {
 
 import { useState } from "react";
 
-import FillInTheBlanks from "@/app/components/challenges/FillInTheBlanks/FillInTheBlanks";
+import FillInTheBlanks from "@/app/components/questions/FillInTheBlanks/FillInTheBlanks";
 import type {
   FillInTheBlanksAttemptResponse,
   FillInTheBlanksQuestion,
